fix: cleanup code and move into seperate components
